[Document] Make sure hardlinks do not delete children of their source (#6025)
* [Document] Make sure hardlinks do not delete children of their source

* [Document] Make sure hardlinks do not delete children of their source

+++ models/Document.php
@@ -23,6 +23,7 @@ use Pimcore\Event\DocumentEvents;
 use Pimcore\Event\FrontendEvents;
 use Pimcore\Event\Model\DocumentEvent;
 use Pimcore\Logger;
+use Pimcore\Model\Document\Hardlink;
 use Pimcore\Model\Document\Hardlink\Wrapper\WrapperInterface;
 use Pimcore\Model\Document\Listing;
 use Pimcore\Model\Element\ElementInterface;
@@ -796,7 +797,9 @@ class Document extends Element\AbstractElement
                 $unpublishedStatus = self::doHideUnpublished();
                 self::setHideUnpublished(false);
                 foreach ($this->getChildren(true) as $child) {
-                    $child->delete(true);
+                    if(!$child instanceof WrapperInterface) {
+                        $child->delete(true);
+                    }
                 }
                 self::setHideUnpublished($unpublishedStatus);
             }
